[FEAT] #142 chatting - 소켓통신 코드 viewmodel에서 activity로 이동~
